refactor(app, shared-data): move splitLabwareDefURI to shared-data
